Crate gstreamer_player

Modules

prelude

Structs

Continue
Continue calling the closure in the future iterations or drop it.

Error
A generic error capable of representing various error domains (types).

Player
Implements

PlayerAudioInfo
PlayerStreamInfo specific to audio streams.

PlayerConfig

PlayerGMainContextSignalDispatcher
Implements

PlayerMediaInfo
Structure containing the media information of a URI.

PlayerSignalDispatcher
Implements

PlayerStreamInfo
Base structure for information concering a media stream. Depending on the stream type, one can find more media-specific information in PlayerVideoInfo, PlayerAudioInfo, PlayerSubtitleInfo.

PlayerSubtitleInfo
PlayerStreamInfo specific to subtitle streams.

PlayerVideoInfo
PlayerStreamInfo specific to video streams.

PlayerVideoOverlayVideoRenderer
Implements

PlayerVideoRenderer
Implements

PlayerVisualization
A PlayerVisualization descriptor.

TypedValue
A statically typed Value.

Value
A generic value capable of carrying various types.

Enums

PlayerColorBalanceType

PlayerError

PlayerSnapshotFormat

PlayerState

Type
A GLib or GLib-based library type

Traits

Cast
Upcasting and downcasting support.

IsA
Declares the "is a" relationship.

PlayerSignalDispatcherExt
Trait containing all PlayerSignalDispatcher methods.

PlayerStreamInfoExt
Trait containing all PlayerStreamInfo methods.

PlayerVideoRendererExt
Trait containing all PlayerVideoRenderer methods.

StaticType
Types that are supported by GLib dynamic typing.

ToValue
Converts to Value.