Struct gstreamer_player::Player

pub struct Player(_, _);

Implements

ObjectExt

Methods

impl Player

pub fn config_set_seek_accurate(&self, accurate: bool)

Enable or disable accurate seeking. When enabled, elements will try harder to seek as accurately as possible to the requested seek position. Generally it will be slower especially for formats that don't have any indexes or timestamp markers in the stream.

If accurate seeking is disabled, elements will seek as close as the request position without slowing down seeking too much.

Accurate seeking is disabled by default.

accurate

accurate seek or not

pub fn get_audio_video_offset(&self) -> i64

Retrieve the current value of audio-video-offset property

Returns

The current value of audio-video-offset in nanoseconds

Since 1.10

pub fn get_color_balance(&self, type_: PlayerColorBalanceType) -> f64

Retrieve the current value of the indicated type_.

type_

PlayerColorBalanceType

Returns

The current value of type_, between [0,1]. In case of error -1 is returned.

pub fn get_config(&self) -> Option<Structure>

Get a copy of the current configuration of the player. This configuration can either be modified and used for the Player::set_config call or it must be freed after usage.

Returns

a copy of the current configuration of self. Use gst::Structure::free after usage or Player::set_config.

Since 1.10

pub fn get_current_audio_track(&self) -> Option<PlayerAudioInfo>

A Function to get current audio PlayerAudioInfo instance.

Returns

current audio track.

The caller should free it with gobject::ObjectExt::unref

pub fn get_current_subtitle_track(&self) -> Option<PlayerSubtitleInfo>

A Function to get current subtitle PlayerSubtitleInfo instance.

Returns

current subtitle track.

The caller should free it with gobject::ObjectExt::unref

pub fn get_current_video_track(&self) -> Option<PlayerVideoInfo>

A Function to get current video PlayerVideoInfo instance.

Returns

current video track.

The caller should free it with gobject::ObjectExt::unref

pub fn get_current_visualization(&self) -> Option<String>

Returns

Name of the currently enabled visualization. g_free after usage.

pub fn get_duration(&self) -> ClockTime

Retrieves the duration of the media stream that self represents.

Returns

the duration of the currently-playing media stream, in nanoseconds.

pub fn get_media_info(&self) -> Option<PlayerMediaInfo>

A Function to get the current media info PlayerMediaInfo instance.

Returns

media info instance.

The caller should free it with gobject::ObjectExt::unref

pub fn get_mute(&self) -> bool

Returns

true if the currently-playing stream is muted.

pub fn get_pipeline(&self) -> Option<Element>

Returns

The internal playbin instance

pub fn get_position(&self) -> ClockTime

Returns

the absolute position time, in nanoseconds, of the currently-playing stream.

pub fn get_rate(&self) -> f64

Returns

current playback rate

pub fn get_subtitle_uri(&self) -> Option<String>

current subtitle URI

Returns

URI of the current external subtitle. g_free after usage.

pub fn get_uri(&self) -> Option<String>

Gets the URI of the currently-playing stream.

Returns

a string containing the URI of the currently-playing stream. g_free after usage.

pub fn get_video_snapshot<'a, P: Into<Option<&'a Structure>>>(
&self,
format: PlayerSnapshotFormat,
config: P
) -> Option<Sample>

Get a snapshot of the currently selected video stream, if any. The format can be selected with format and optional configuration is possible with config Currently supported settings are: - width, height of type G_TYPE_INT - pixel-aspect-ratio of type GST_TYPE_FRACTION Except for GST_PLAYER_THUMBNAIL_RAW_NATIVE format, if no config is set, pixel-aspect-ratio would be 1/1

format

output format of the video snapshot

config

Additional configuration

Returns

Current video snapshot sample or None on failure

Since 1.12

pub fn get_volume(&self) -> f64

Returns the current volume level, as a percentage between 0 and 1.

Returns

the volume as percentage between 0 and 1.

pub fn has_color_balance(&self) -> bool

Checks whether the self has color balance support available.

Returns

true if self has color balance support. Otherwise, false.

pub fn pause(&self)

Pauses the current stream.

pub fn play(&self)

Request to play the loaded stream.

pub fn seek(&self, position: ClockTime)

Seeks the currently-playing stream to the absolute position time in nanoseconds.

position

position to seek in nanoseconds

pub fn set_audio_track(&self, stream_index: i32) -> bool

stream_index

stream index

Returns

true or false

Sets the audio track stream_idex.

pub fn set_audio_track_enabled(&self, enabled: bool)

Enable or disable the current audio track.

enabled

TRUE or FALSE

pub fn set_audio_video_offset(&self, offset: i64)

Sets audio-video-offset property by value of offset

Since 1.10

offset

gint64 in nanoseconds

pub fn set_color_balance(&self, type_: PlayerColorBalanceType, value: f64)

Sets the current value of the indicated channel type_ to the passed value.

type_

PlayerColorBalanceType

value

The new value for the type_, ranged [0,1]

pub fn set_mute(&self, val: bool)

true if the currently-playing stream should be muted.

val

Mute state the should be set

pub fn set_rate(&self, rate: f64)

Playback at specified rate

rate

playback rate

pub fn set_subtitle_track(&self, stream_index: i32) -> bool

stream_index

stream index

Returns

true or false

Sets the subtitle strack stream_index.

pub fn set_subtitle_track_enabled(&self, enabled: bool)

Enable or disable the current subtitle track.

enabled

TRUE or FALSE

pub fn set_subtitle_uri(&self, uri: &str)

Sets the external subtitle URI.

uri

subtitle URI

pub fn set_uri(&self, uri: &str)

Sets the next URI to play.

uri

next URI to play.

pub fn set_video_track(&self, stream_index: i32) -> bool

stream_index

stream index

Returns

true or false

Sets the video track stream_index.

pub fn set_video_track_enabled(&self, enabled: bool)

Enable or disable the current video track.

enabled

TRUE or FALSE

pub fn set_visualization(&self, name: &str) -> bool

name

visualization element obtained from Player::visualizations_get()

Returns

true if the visualizations was set correctly. Otherwise, false.

pub fn set_visualization_enabled(&self, enabled: bool)

Enable or disable the visualization.

enabled

TRUE or FALSE

pub fn set_volume(&self, val: f64)

Sets the volume level of the stream as a percentage between 0 and 1.

val

the new volume level, as a percentage between 0 and 1

pub fn stop(&self)

Stops playing the current stream and resets to the first position in the stream.

pub fn get_property_suburi(&self) -> Option<String>

pub fn set_property_suburi(&self, suburi: Option<&str>)

pub fn config_get_position_update_interval(config: &Structure) -> u32

config

a Player configuration

Returns

current position update interval in milliseconds

Since 1.10

pub fn config_get_seek_accurate(config: &Structure) -> bool

config

a Player configuration

Returns

true if accurate seeking is enabled

Since 1.12

pub fn config_get_user_agent(config: &Structure) -> Option<String>

Return the user agent which has been configured using Player::config_set_user_agent if any.

config

a Player configuration

Returns

the configured agent, or None Since 1.10

pub fn config_set_position_update_interval(
config: &mut Structure,
interval: u32
)

set interval in milliseconds between two position-updated signals. pass 0 to stop updating the position. Since 1.10

config

a Player configuration

interval

interval in ms

pub fn config_set_user_agent(config: &mut Structure, agent: &str)

Set the user agent to pass to the server if player needs to connect to a server during playback. This is typically used when playing HTTP or RTSP streams.

Since 1.10

config

a Player configuration

agent

the string to use as user agent

pub fn get_audio_streams(info: &PlayerMediaInfo) -> Vec<PlayerAudioInfo>

info

a PlayerMediaInfo

Returns

A glib::List of matching PlayerAudioInfo.

pub fn get_subtitle_streams(info: &PlayerMediaInfo) -> Vec<PlayerSubtitleInfo>

info

a PlayerMediaInfo

Returns

A glib::List of matching PlayerSubtitleInfo.

pub fn get_video_streams(info: &PlayerMediaInfo) -> Vec<PlayerVideoInfo>

info

a PlayerMediaInfo

Returns

A glib::List of matching PlayerVideoInfo.

pub fn visualizations_get() -> Vec<PlayerVisualization>

Returns

a None terminated array containing all available visualizations. Use Player::visualizations_free after usage.

pub fn connect_buffering<F: Fn(&Player, i32) + Send + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_duration_changed<F: Fn(&Player, u64) + Send + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_end_of_stream<F: Fn(&Player) + Send + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_error<F: Fn(&Player, &Error) + Send + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_media_info_updated<F: Fn(&Player, &PlayerMediaInfo) + Send + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_mute_changed<F: Fn(&Player) + Send + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_position_updated<F: Fn(&Player, u64) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_seek_done<F: Fn(&Player, u64) + Send + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_state_changed<F: Fn(&Player, PlayerState) + Send + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_uri_loaded<F: Fn(&Player, &str) + Send + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_video_dimensions_changed<F: Fn(&Player, i32, i32) + Send + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_volume_changed<F: Fn(&Player) + Send + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_warning<F: Fn(&Player, &Error) + Send + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_audio_video_offset_notify<F: Fn(&Player) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_current_audio_track_notify<F: Fn(&Player) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_current_subtitle_track_notify<F: Fn(&Player) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_current_video_track_notify<F: Fn(&Player) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_duration_notify<F: Fn(&Player) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_media_info_notify<F: Fn(&Player) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_mute_notify<F: Fn(&Player) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_pipeline_notify<F: Fn(&Player) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_position_notify<F: Fn(&Player) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_rate_notify<F: Fn(&Player) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_signal_dispatcher_notify<F: Fn(&Player) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_suburi_notify<F: Fn(&Player) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_uri_notify<F: Fn(&Player) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_video_multiview_flags_notify<F: Fn(&Player) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_video_multiview_mode_notify<F: Fn(&Player) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_video_renderer_notify<F: Fn(&Player) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_volume_notify<F: Fn(&Player) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

impl Player

pub fn new(
video_renderer: Option<&PlayerVideoRenderer>,
signal_dispatcher: Option<&PlayerSignalDispatcher>
) -> Player

Creates a new Player instance that uses signal_dispatcher to dispatch signals to some event loop system, or emits signals directly if NULL is passed. See PlayerGMainContextSignalDispatcher::new.

Video is going to be rendered by video_renderer, or if None is provided no special video set up will be done and some default handling will be performed.

video_renderer

GstPlayerVideoRenderer to use

signal_dispatcher

GstPlayerSignalDispatcher to use

Returns

a new Player instance

pub fn set_config(&self, config: Structure) -> bool