Struct gstreamer_player::PlayerVideoOverlayVideoRenderer

pub struct PlayerVideoOverlayVideoRenderer(_, _);

Implements

ObjectExt, PlayerVideoRendererExt

Methods

impl PlayerVideoOverlayVideoRenderer

fn expose(&self)

Tell an overlay that it has been exposed. This will redraw the current frame in the drawable even if the pipeline is PAUSED.

fn get_render_rectangle(&self) -> (i32, i32, i32, i32)

Return the currently configured render rectangle. See PlayerVideoOverlayVideoRenderer::set_render_rectangle for details.

x

the horizontal offset of the render area inside the window

y

the vertical offset of the render area inside the window

width

the width of the render area inside the window

height

the height of the render area inside the window

fn set_render_rectangle(&self, x: i32, y: i32, width: i32, height: i32)

Configure a subregion as a video target within the window set by PlayerVideoOverlayVideoRenderer::set_window_handle. If this is not used or not supported the video will fill the area of the window set as the overlay to 100%. By specifying the rectangle, the video can be overlaid to a specific region of that window only. After setting the new rectangle one should call PlayerVideoOverlayVideoRenderer::expose to force a redraw. To unset the region pass -1 for the width and height parameters.

This method is needed for non fullscreen video overlay in UI toolkits that do not support subwindows.

x

the horizontal offset of the render area inside the window

y

the vertical offset of the render area inside the window

width

the width of the render area inside the window

height

the height of the render area inside the window

fn get_property_video_sink(&self) -> Option<Element>

fn set_property_video_sink<P: IsA<Element> + IsA<Object> + SetValueOptional>(
&self,
video_sink: Option<&P>
)

fn connect_property_video_sink_notify<F: Fn(&PlayerVideoOverlayVideoRenderer) + Send + Sync + 'static>(
&self,
f: F
) -> u64

fn connect_property_window_handle_notify<F: Fn(&PlayerVideoOverlayVideoRenderer) + Send + Sync + 'static>(
&self,
f: F
) -> u64

impl PlayerVideoOverlayVideoRenderer

unsafe fn new(window_handle: uintptr_t) -> PlayerVideoOverlayVideoRenderer

window_handle

Window handle to use or None

unsafe fn new_with_handle_and_sink<P: IsA<Element>>(
window_handle: uintptr_t,
video_sink: &P
) -> PlayerVideoOverlayVideoRenderer

fn new_with_sink<P: IsA<Element>>(
video_sink: &P
) -> PlayerVideoOverlayVideoRenderer

window_handle

Window handle to use or None

video_sink

the custom video_sink element to be set for the video renderer

Returns

Since 1.12